Use 4-channel pixel for font texture atlas

+++ bootstrap/include/std/font_provider.h
@@ -15,7 +15,7 @@ STRUCT(FontCharacter)
 
 STRUCT(TextureAtlas)
 {
-    u8* pointer;
+    u32* pointer;
     FontCharacter* characters;
     s32* kerning_tables;
     u32 width;
